% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniTidyR-package.R
\docType{package}
\name{miniTidyR-package}
\alias{miniTidyR}
\alias{miniTidyR-package}
\title{miniTidyR: Data wrangling made easier}
\description{
Dataframes are fun to work with but they do not always come in the desired shape.
          The tidyverse functions 'gather', 'spread' and 'drop_na' gracefully address some of the messy dataframe issues. 
          To better understand these functddions, we re-implement them in this package with the following names: my_gather, my_spread, my_dropna.
}
\author{
\strong{Maintainer}: first last \email{first.last@example.com}

}
\keyword{internal}
