% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_dropna.R
\name{my_dropna}
\alias{my_dropna}
\title{Returns a new dataframe with the entire rows that contain `na` values removed}
\usage{
my_dropna(input_dataframe)
}
\arguments{
\item{input_dataframe}{data.frame}
}
\value{
data.frame
}
\description{
Returns a new dataframe with the entire rows that contain `na` values removed
}
\examples{
my_dropna(input_dataframe)
}
