% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_gather.R
\name{my_gather}
\alias{my_gather}
\title{Tranform a dataframe from wide to long}
\usage{
my_gather(data, key = NULL, value = NULL, chosen_cols = NULL)
}
\arguments{
\item{data}{data.frame}

\item{key}{numeric or character}

\item{value}{numeric or character}

\item{chosen_cols}{numeric or character}
}
\value{
data.frame
}
\description{
Tranform a dataframe from wide to long
}
\examples{
df <- data.frame(Groups = c("A","B"), Married = c(12,25), Other = c(100,55))
my_gather(df, status,count,Married,Other)

}
