% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_spread.R
\name{my_spread}
\alias{my_spread}
\title{Transform a dataframe from long to wide}
\usage{
my_spread(data, keycol, valcol)
}
\arguments{
\item{data}{A data frame}

\item{keycol}{Name of column whose content will be the names of the new columns, a string}

\item{valcol}{Name of column whose content will be the values of the new columns, a sting}
}
\value{
A data frame
}
\description{
Transform a dataframe from long to wide
}
\examples{
df <- data.frame(Name = c("A","B"), key = c("Age","Weight"), value = c(45,80))
my_spread(df,"key","value")


}
