% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_spread.R
\name{my_spread}
\alias{my_spread}
\title{Transform a dataframe from long to wide}
\usage{
my_spread(data, key, value)
}
\arguments{
\item{data}{data.frame}

\item{key}{numeric or character}

\item{value}{numeric or character}
}
\value{
data.frame
}
\description{
Transform a dataframe from long to wide
}
\examples{
df <- data.frame(Name = c("A","A"), key = c("Age","Weight"), value = c(45,80))
my_spread(df,key,value)

}
